Fix php code
